Lots of changes, see description
1. Added blobstore script handler to app.yaml
2. Added file upload handlers to project settings (settings.py)
3. Added blobstore handler file for uploading, removing, serving and
creating upload urls.
4. Added comments to categories.py
5. Added controller code to menu.py
6. Web api for categories was refactored
7. Web api for items was refactored
8. Web api for pings was refactored
9. Web api for orders was refactored
10. Web api for recommendations was refactored
11. Web api for tables was refactored
12. Added helpers.py for general methods that help on data processing
13. menu.html reviewed
14. new.html for new items reviewed
15. lists.py didn’t changed
16. testItem.py didn’t changed
17. Added file for testOrder, testOrderDetail,  testPing and
testRecommendation
18. Deleted todoTests file
19. Updated urls.py, now is more homogeneous
20. Removed url from default drop zone values, set uploadMultiple to
false, and bumped maxFilesize from 100 to 256, the autoProcessQueue
proper tie is !1 = false, and the plugin doesn’t auto discover any drop
zone anymore
21. Endless wizard now handles a submit button and hidden input types
(for files)
22. Parsley was de-minified

+++ lobsternachos/lobsternachos/controllers/categories.py
@@ -2,29 +2,48 @@ from django.http import HttpResponseRedirect
 from lobsternachos.models import *
 from google.appengine.ext import ndb
 import urllib
+from lobsternachos.helpers import *
 
 def create(request):
   if request.method == 'POST':
+    # Create empty category
     category = Category(parent=GetAncestor())
+
+    # If there exist a name parameter
     if request.POST.get('categoryName'):
+      # Assign and include
       category.CategoryName=request.POST.get('categoryName')
       categoryID = category.put().integer_id()
+
+      # Redirect with category id
       return HttpResponseRedirect('/menu?' +  urllib.urlencode({'categoryID':categoryID}))
   return HttpResponseRedirect('/menu')
 
 def delete(request):
   if request.method == 'POST':
-    category = Category.get_by_id(int(request.POST.get('categoryID')),
-    parent = GetAncestor())
-    category.key.delete()
+    # Check if category id is long
+    if isLong( request.POST.get('categoryID') ):
+      # Try to get category
+      category = Category.get_by_id(long(request.POST.get('categoryID')),
+      parent = GetAncestor())
+      # Check if exist
+      if category:
+        category.key.delete()
   return HttpResponseRedirect('/menu')
 
 def update(request):
   if request.method == 'POST':
-    category = Category.get_by_id(int(request.POST.get('categoryID')),
-    parent = GetAncestor())
-    if category and request.POST.get('categoryName'):
-      category.CategoryName = request.POST.get('categoryName')
-      categoryID = category.put().integer_id()
-      return HttpResponseRedirect('/menu?' +  urllib.urlencode({'categoryID':categoryID}))
+    # Check if category id is long
+    if isLong( request.POST.get('categoryID') ):
+      # Try to get category
+      category = Category.get_by_id(long(request.POST.get('categoryID')),
+      parent = GetAncestor())
+      # If exist and category name is present
+      if category and request.POST.get('categoryName'):
+        # Update
+        category.CategoryName = request.POST.get('categoryName')
+        categoryID = category.put().integer_id()
+
+        #Redirect with category id
+        return HttpResponseRedirect('/menu?' +  urllib.urlencode({'categoryID':categoryID}))
   return HttpResponseRedirect('/menu')
